Add Me fragment
Also broke a couple things, including the slack button.
