+++ app/src/main/res/layout/fragment_main_screen.xml
@@ -47,6 +47,7 @@
             android:layout_height="wrap_content"
             android:adjustViewBounds="true"
             android:cropToPadding="true"
+            android:layout_below="@id/attendance_layout"
             android:paddingBottom="@dimen/activity_vertical_margin"
             android:scaleType="fitXY"/>
 
@@ -77,41 +78,10 @@
 
 
             <TextView
-                android:id="@+id/slack_textview"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_below="@id/about_text_view"
-                android:paddingTop="@dimen/activity_vertical_margin"
-                android:visibility="gone"/>
-
-            <Button
-                android:id="@+id/slack_sign_up_button"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_alignTop="@id/slack_textview"
-                android:layout_toRightOf="@id/slack_textview"
-                android:visibility="gone"/>
-
-            <TextView
-                android:id="@+id/current_committee_textview"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_below="@id/slack_sign_up_button"
-                android:text="Choose a committee"/>
-
-            <Button
-                android:id="@+id/change_committee_button"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_alignTop="@id/current_committee_textview"
-                android:layout_toRightOf="@id/current_committee_textview"
-                android:text="Change committee"/>
-
-            <TextView
                 android:id="@+id/google_sign_out_textview"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:layout_below="@id/change_committee_button"
+                android:layout_below="@id/about_text_view"
                 android:paddingTop="@dimen/activity_vertical_margin"
                 android:visibility="gone"
                 android:text="@string/you_are_signed_into_your_google_account"/>
